first upload
this code is based on that of my own Hacker News Amplifier (ver. 1.6),
the Chrome extension published at
https://chrome.google.com/webstore/detail/jaapjannjfdcklcgadbhiddhbelhhn
kp

this code has two important changes: (1) a refactoring that makes the
extension easier to adapt for lists different from Hacker News, like
Reddit; and (2) all the needed code for working on Reddit too.